Introduce stdbool.h

+++ src/drivers/serial.h
@@ -24,8 +24,8 @@
 #define SERIAL_SPEED_38400  3
 
 void serial_init(uint16_t com, uint16_t divisor);
-void serial_print(uint32_t com, const char* str);
-void serial_write(uint32_t com, char c);
-void serial_printf(uint32_t com, const char* format, ...);
+void serial_print(uint16_t com, const char* str);
+void serial_write(uint16_t com, char c);
+void serial_printf(uint16_t com, const char* format, ...);
 
 #endif
